+++ CHANGELOG.md
@@ -3,6 +3,19 @@ All notable changes to this project will be documented in this file.
 
 Note that Objective-C class names are prefixed by `SNR`. In the changelog below, these are names used in Swift, without the prefix.
 
+## [4.15.0] - 2024-03-06
+
+### Added
+- Global Control Group support for in-app messages. From now on, you can use this feature in in-app messaging communication. This lets you take your marketing efforts to the next level and provides a solid foundation for accurate measurement of campaign effectiveness. Read more at https://hub.synerise.com/docs/settings/configuration/global-control-group/.
+- We added a new `Client.authenticate(tokenPayload:authID:success:failure:)` method. This method signs in a customer in with the provided token payload.
+- We added a new `Client.getUUIDForAuthentication(authID:)` method. This method retrieves the current UUID or generates a new one from a seed (`authId`).
+- `clientId` property in the `Token` model.
+
+### Changed
+- We added additional validation for the `Synerise.initialize(clientApiKey:)` and `Synerise.initialize(clientApiKey:baseUrl:)` methods. Now, if you try to initialize the sdk when it's already running, the method is blocked and no action is invoked.
+- Improvements to stability.
+
+
 ## [4.14.12] - 2024-02-27
 
 ### Fixed